move error_report to ErrorCodes.cpp
