+++ rpcs3/Emu/System.h
@@ -468,8 +468,6 @@ public:
 
 extern Emulator Emu;
 
-extern bool g_log_all_errors;
-
 extern bool g_use_rtm;
 extern u64 g_rtm_tx_limit1;
 extern u64 g_rtm_tx_limit2;
